Refactored all modules.
